Refs #13008: Fix test broken in rebase
