+++ test/unittest/participant/participant_factory/ParticipantFactoryTest.cpp
@@ -21,6 +21,7 @@
 #include <ddsrouter/exceptions/ConfigurationException.hpp>
 #include <ddsrouter/participant/IParticipant.hpp>
 #include <ddsrouter/participant/ParticipantFactory.hpp>
+#include <ddsrouter/types/configuration_tags.hpp>
 #include <ddsrouter/types/participant/ParticipantId.hpp>
 #include <ddsrouter/types/participant/ParticipantType.hpp>
 
@@ -30,18 +31,51 @@ namespace eprosima {
 namespace ddsrouter {
 namespace test {
 
+// TODO: refactor these tests so they test with type in id and with type in Configuration
+// TODO: add these functions in a common utils test
+void set_domain(
+        RawConfiguration& configuration,
+        uint16_t seed = 0)
+{
+    configuration[DOMAIN_ID_TAG] = seed;
+}
+
+ParticipantConfiguration get_random_participant_configuration(
+    ParticipantId id,
+    ParticipantType type,
+    uint32_t seed = 0)
+{
+    RawConfiguration config;
+
+    switch (type())
+    {
+        case ParticipantType::SIMPLE_RTPS:
+            set_domain(config, seed);
+            break;
+
+        // Add configurations por Participant Types that require configuration arguments
+
+        default:
+            break;
+    }
+
+    return ParticipantConfiguration(id, config);
+}
+
 /*
  * Generate all required objects for participant creation,
  * and then create a participant from an id string
  */
 std::shared_ptr<IParticipant> create_participant(
-        std::string id_str)
+        std::string id_str,
+        ParticipantType type)
 {
     ParticipantFactory participant_factory;
     std::shared_ptr<PayloadPool> payload_pool = std::make_shared<PayloadPool>();
     std::shared_ptr<DiscoveryDatabase> discovery_database = std::make_shared<DiscoveryDatabase>();
     ParticipantId id(id_str);
-    ParticipantConfiguration participant_configuration = ParticipantConfiguration(id);
+    ParticipantConfiguration participant_configuration =
+        get_random_participant_configuration(id, type);
 
     return participant_factory.create_participant(participant_configuration, payload_pool, discovery_database);
 }
@@ -58,7 +92,7 @@ std::shared_ptr<IParticipant> create_participant(
  */
 TEST(ParticipantFactoryTest, create_void_participant)
 {
-    std::shared_ptr<IParticipant> void_participant = test::create_participant("void");
+    std::shared_ptr<IParticipant> void_participant = test::create_participant("void", ParticipantType::VOID);
     ASSERT_EQ(void_participant->type()(), ParticipantType::VOID);
 }
 
@@ -70,7 +104,7 @@ TEST(ParticipantFactoryTest, create_void_participant)
  */
 TEST(ParticipantFactoryTest, create_echo_participant)
 {
-    std::shared_ptr<IParticipant> echo_participant = test::create_participant("echo");
+    std::shared_ptr<IParticipant> echo_participant = test::create_participant("echo", ParticipantType::ECHO);
     ASSERT_EQ(echo_participant->type()(), ParticipantType::ECHO);
 }
 
@@ -82,7 +116,7 @@ TEST(ParticipantFactoryTest, create_echo_participant)
  */
 TEST(ParticipantFactoryTest, create_dummy_participant)
 {
-    std::shared_ptr<IParticipant> dummy_participant = test::create_participant("dummy");
+    std::shared_ptr<IParticipant> dummy_participant = test::create_participant("dummy", ParticipantType::DUMMY);
     ASSERT_EQ(dummy_participant->type()(), ParticipantType::DUMMY);
 }
 
@@ -94,7 +128,7 @@ TEST(ParticipantFactoryTest, create_dummy_participant)
  */
 TEST(ParticipantFactoryTest, create_simple_participant)
 {
-    std::shared_ptr<IParticipant> simple_participant = test::create_participant("local");
+    std::shared_ptr<IParticipant> simple_participant = test::create_participant("local", ParticipantType::SIMPLE_RTPS);
     ASSERT_EQ(simple_participant->type()(), ParticipantType::SIMPLE_RTPS);
 }
 
@@ -106,7 +140,7 @@ TEST(ParticipantFactoryTest, create_simple_participant)
  */
 TEST(ParticipantFactoryTest, create_invalid_participant)
 {
-    ASSERT_THROW(test::create_participant("invalid_part"), ConfigurationException);
+    ASSERT_THROW(test::create_participant("invalid_part", ParticipantType::VOID), ConfigurationException);
 }
 
 /**
